+++ tests/integration/cases/scalar_int_very_large/OCaml.ml
@@ -2,6 +2,6 @@ module Check = struct
 
 type val_t =
   | OInt of int
-let my_data : val_t = OInt (int_of_string "9223372036854775808")
+let my_data : val_t = OInt (try int_of_string "9223372036854775808" with Failure _ -> 0)
 
 end